+++ Tests/BugsnagTests/BugsnagEventTests.m
@@ -18,7 +18,6 @@
 #import "BugsnagSession.h"
 #import "BugsnagSession+Private.h"
 #import "BugsnagStackframe+Private.h"
-#import "BugsnagStateEvent.h"
 #import "BugsnagTestConstants.h"
 #import "BugsnagTestsDummyClass.h"
 
@@ -321,6 +320,36 @@
         NSDictionary *toJson = [event toJsonWithRedactedKeys:nil];
         XCTAssertEqualObjects(json, toJson, @"Input and output JSON do not match");
     }
+}
+
+// MARK: - Feature flags interface
+
+- (void)testFeatureFlags {
+    BugsnagEvent *event = [[BugsnagEvent alloc] initWithKSReport:@{
+        @"user.metaData": @{
+                @"user": @{@"id": @"user id"}
+        }}];
+    
+    XCTAssertEqualObjects([event toJsonWithRedactedKeys:nil][@"featureFlags"], @[]);
+    
+    [event addFeatureFlagWithName:@"color" variant:@"red"];
+    
+    XCTAssertEqualObjects([event toJsonWithRedactedKeys:nil][@"featureFlags"],
+                          (@[@{@"featureFlag": @"color", @"variant": @"red"}]));
+    
+    [event addFeatureFlagWithName:@"color" variant:@"green"];
+    
+    XCTAssertEqualObjects([event toJsonWithRedactedKeys:nil][@"featureFlags"],
+                          (@[@{@"featureFlag": @"color", @"variant": @"green"}]));
+    
+    [event addFeatureFlagWithName:@"color"];
+    
+    XCTAssertEqualObjects([event toJsonWithRedactedKeys:nil][@"featureFlags"],
+                          (@[@{@"featureFlag": @"color"}]));
+    
+    [event clearFeatureFlags];
+    
+    XCTAssertEqualObjects([event toJsonWithRedactedKeys:nil][@"featureFlags"], @[]);
 }
 
 // MARK: - Metadata interface